api_test_framework/requirements.txt
pytest==8.1.1
pytest-html==4.1.1
allure-pytest==2.13.5
requests==2.31.0
python-dotenv==1.0.1
jsonschema==4.22.0
colorlog==6.8.2
rich==13.7.1
pyyaml==6.0.1
faker==24.11.0
pytest-xdist==3.5.0
pytest-rerunfailures==14.0
